# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.categoryListResp:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handler.categoryResp:
    properties:
      data:
        $ref: '#/definitions/model.Category'
    type: object
  handler.errResp:
    properties:
      error:
        type: string
    type: object
...
  model.Category:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.CreateCategoryRequest:
    properties:
      color:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  model.UpdateCategoryRequest:
    properties:
      color:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for TodoList — manage todos, categories, and user accounts.
  title: TodoList API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Return all categories for the authenticated user.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.categoryListResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errResp'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: 'Create a new category. Color defaults to #6366f1 if not provided.'
      parameters:
      - description: Category payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.categoryResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errResp'
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Permanently delete a category.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errResp'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - categories
    get:
      description: Return a single category by ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.categoryResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errResp'
      security:
      - BearerAuth: []
      summary: Get a category
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update a category's name or color.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.categoryResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errResp'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"